refactor(sample): register the live notification renderer from Application
Both apps registered the custom renderer in MainActivity.onCreate, which does not
run when Android starts a process solely to deliver a push — the case that
matters most, since that is when a pushed live notification needs rendering.
Application.onCreate runs in every process, so move it there.

The renderer was also duplicated verbatim in both apps. It now lives once in the
shared Android source set, alongside a shared Application that both manifests
name in full, replacing Flutter's ${applicationName} placeholder (which resolves
to android.app.Application, the class extended here). This matches the React
Native sample, which already registers from MainApplication.
