+++ apps/flutter_sample_cocoapods/android/app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
 
     <application
-        android:name="${applicationName}"
+        android:name="io.customer.testbed.flutter.MainApplication"
         android:icon="@mipmap/ic_launcher"
         android:label="Flutter Pods">
         <activity
