+++ apps/flutter_sample_cocoapods/android/app/src/main/kotlin/io/customer/testbed/flutter/cocoapods/MainActivity.kt
@@ -1,15 +1,5 @@
 package io.customer.testbed.flutter.cocoapods
 
-import android.app.Notification
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
-import android.os.Build
-import android.os.Bundle
-import androidx.core.app.NotificationCompat
-import io.customer.customer_io.liveactivities.CustomerIOLiveActivities
-import io.customer.messagingpush.data.communication.CustomerIOLiveNotificationsCallback
-import io.customer.messagingpush.data.model.CustomerIOParsedPushPayload
 import io.customer.testbed.flutter.PermissionChannelHandler
 import io.flutter.embedding.android.FlutterActivity
 import io.flutter.embedding.engine.FlutterEngine
@@ -17,15 +7,6 @@ import io.flutter.plugin.common.MethodChannel
 
 class MainActivity : FlutterActivity() {
     private lateinit var permissionHandler: PermissionChannelHandler
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        // Custom live notifications require a render callback that can only be set at SDK build time.
-        // Register it BEFORE the Dart layer initializes the Customer.io SDK (which happens later,
-        // once the Flutter engine runs). The wrapper stores this statically and applies it onto the
-        // push module config during SDK init.
-        CustomerIOLiveActivities.setLiveNotificationCallback(RideshareLiveNotificationCallback())
-        super.onCreate(savedInstanceState)
-    }
 
     override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
         super.configureFlutterEngine(flutterEngine)
@@ -39,70 +20,5 @@ class MainActivity : FlutterActivity() {
         if (!permissionHandler.onRequestPermissionsResult(requestCode, permissions, grantResults)) {
             super.onRequestPermissionsResult(requestCode, permissions, grantResults)
         }
-    }
-}
-
-/**
- * Sample app-owned renderer for the custom `rideshare` live notification type. Custom types have no
- * built-in SDK template, so the app fully renders them here. Returning null for any other type lets
- * the SDK fall back to its built-in templates.
- *
- * The `data` map Dart sends with `LiveActivityPayload.custom` arrives flattened in
- * [CustomerIOParsedPushPayload.extras], keyed the same way.
- */
-private class RideshareLiveNotificationCallback : CustomerIOLiveNotificationsCallback {
-    override fun createLiveNotification(
-        payload: CustomerIOParsedPushPayload,
-        context: Context,
-    ): Notification? {
-        val extras = payload.extras
-        if (extras.getString("notification_type") != RIDESHARE_TYPE) return null
-
-        // The SDK re-invokes this callback on the "end" event. Return a terminal, non-ongoing
-        // notification then so it can be dismissed instead of sticking around forever.
-        val ended = extras.getString("event") == "end"
-
-        val driverName = extras.getString("driverName") ?: "Your driver"
-        val status = extras.getString("status") ?: ""
-        // Every custom value is a string, here and in the pushed content-state; parse what you need.
-        val etaMinutes = extras.getString("etaMinutes")?.toDoubleOrNull()?.toInt()
-
-        // Mirrors the iOS widget's layout: status leads (as the title below), then the driver, then
-        // the ETA. `status` is not repeated here since it is already the title.
-        val body = listOfNotNull(
-            driverName,
-            etaMinutes?.let { "ETA $it min" },
-        ).joinToString(" • ")
-
-        ensureChannel(context)
-        return NotificationCompat.Builder(context, CHANNEL_ID)
-            .setSmallIcon(context.applicationInfo.icon)
-            // Title follows `status`, matching the iOS widget's headline. Hardcoding "is on the way"
-            // left Android showing stale copy after an update that changed the status, while the
-            // body below already reflected the new state.
-            .setContentTitle(status.ifEmpty { if (ended) "Arrived" else "Rideshare" })
-            .setContentText(body)
-            .setOngoing(!ended)
-            .setOnlyAlertOnce(true)
-            .build()
-    }
-
-    private fun ensureChannel(context: Context) {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O) return
-        val manager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        if (manager.getNotificationChannel(CHANNEL_ID) == null) {
-            manager.createNotificationChannel(
-                NotificationChannel(
-                    CHANNEL_ID,
-                    "Rideshare",
-                    NotificationManager.IMPORTANCE_DEFAULT,
-                ),
-            )
-        }
-    }
-
-    private companion object {
-        const val RIDESHARE_TYPE = "io.customer.livenotifications.custom.rideshare"
-        const val CHANNEL_ID = "rideshare_live_notification"
     }
 }
